requirements.txt
streamlit
pandas
matplotlib
textblob
wordcloud
google-generativeai
numpy
